refactor: auth 패키지와 shoppingcart 패키지의 Config를 병합
