--- src/main/java/woowacourse/auth/config/AuthenticationPrincipalConfig.java
@@ -1,30 +1,0 @@
-package woowacourse.auth.config;
-
-import java.util.List;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import woowacourse.auth.controller.CustomerIdArgumentResolver;
-import woowacourse.auth.controller.TokenArgumentResolver;
-import woowacourse.auth.support.JwtTokenProvider;
-
-@Configuration
-public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
-    private final JwtTokenProvider jwtTokenProvider;
-
-    public AuthenticationPrincipalConfig(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new TokenArgumentResolver());
-        resolvers.add(userIdArgumentResolver());
-    }
-
-    @Bean
-    public CustomerIdArgumentResolver userIdArgumentResolver() {
-        return new CustomerIdArgumentResolver(jwtTokenProvider);
-    }
-}
